+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Breaking:
   - `RxDocument.putAttachment()` no longer supports string as data, only `Blob` or `Buffer`
   - Pouchdb plugins are not longer added via `addRxPlugin()` but `addPouchPlugin()`. (RxDB plugins are still added via `addRxPlugin`)
+  - Removed plugin hook `preCreatePouchDb`
 
 ### 9.20.0 (15 May 2021)